% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi.amat}
\alias{as_caugi.amat}
\title{Convert a pcalg amat to a \code{caugi_graph}}
\usage{
as_caugi.amat(x, amat_type = NULL, ...)
}
\arguments{
\item{x}{A pcalg adjacency matrix (amat).}

\item{amat_type}{The type of the adjacency matrix, \code{"cpdag"} or \verb{"pag}.}

\item{...}{Additional arguments (will not be passed to anything).}
}
\description{
Convert a pcalg adjacency matrix to a \code{caugi_graph}.
The matrix must be a square matrix with integer edge codes.
See pcalg documentation on amat for more details on the edge codes in amat.
}
