% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi.data.frame}
\alias{as_caugi.data.frame}
\title{Convert a data frame to a \code{caugi_graph}}
\usage{
\method{as_caugi}{data.frame}(x, ...)
}
\arguments{
\item{x}{A data frame with columns \code{from}, \code{to}, and \code{edge_type}.}

\item{...}{Additional arguments (will not be passed to anything).}
}
\description{
Should be a data frame with columns \code{from}, \code{to}, and \code{edge_type}.
}
