% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi.default}
\alias{as_caugi.default}
\title{Default method for \code{as_caugi}}
\usage{
as_caugi.default(x, ...)
}
\arguments{
\item{x}{An object to convert to a \code{caugi_graph}.}

\item{...}{Additional arguments (will not be passed to anything).}
}
\description{
Default method for \code{as_caugi} that throws an error.
}
