% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi.graphNEL}
\alias{as_caugi.graphNEL}
\title{Create a caugi_graph from a possibly‐mixed graphNEL}
\usage{
\method{as_caugi}{graphNEL}(x, collapse = FALSE, collapse_to = "---", ...)
}
\arguments{
\item{x}{A graphNEL object}

\item{collapse}{Logical, whether to collapse directed edges to undirected}

\item{collapse_to}{The code to use for collapsed directed pairs (default "---")}

\item{...}{Additional arguments (will not be passed to anything).}
}
\description{
Create a caugi_graph from a possibly‐mixed graphNEL
}
