% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{as_caugi.igraph}
\alias{as_caugi.igraph}
\title{Convert an \strong{igraph} object to a \code{caugi_graph}}
\usage{
\method{as_caugi}{igraph}(x, collapse = TRUE, ...)
}
\arguments{
\item{x}{An \code{\link{igraph}} object.}

\item{collapse}{Logical. If \code{TRUE} (default) \emph{mutual} pairs
(\verb{A→B} \strong{and} \verb{B→A}) that are marked as symmetric
(\code{edge_type} \%in\% c("---","<->")) are collapsed to
one row.  Set to \code{FALSE} to keep every edge exactly
as it appears in the igraph.}

\item{...}{Not used; kept for S3 compatibility.}
}
\value{
A \code{caugi_graph}.
}
\description{
Convert an \strong{igraph} object to a \code{caugi_graph}
}
