% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi.matrix}
\alias{as_caugi.matrix}
\title{Create a caugi_graph from an integer-coded dense matrix}
\usage{
\method{as_caugi}{matrix}(x, ...)
}
\arguments{
\item{x}{A dense matrix with integer edge codes}

\item{...}{Additional arguments (will not be passed to anything).}
}
\description{
Create a \code{caugi_graph} from a dense matrix with integer edge codes.
The edge codes are:
\itemize{
\item 0: no edge
\item 1: directed edge (from --> to)
\item 2: bidirectional edge (from <-> to)
\item 3: PAG directed edge (from o-> to)
\item 4: PAG undirected edge (from o-- to)
\item 5: PAG bidirected edge (from o-o to)
\item 6: undirected edge (from --- to)
}
}
