% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R, R/caugi_from.R
\name{as_caugi.dgCMatrix}
\alias{as_caugi.dgCMatrix}
\alias{as_caugi.sparseMatrix}
\alias{caugi_from_sparse}
\title{Create a caugi_graph from a sparse \code{dgCMatrix} or \code{sparseMatrix}}
\usage{
\method{as_caugi}{dgCMatrix}(x, directed = TRUE, ...)

\method{as_caugi}{sparseMatrix}(x, directed = TRUE, ...)

caugi_from_sparse(x, directed)
}
\arguments{
\item{x}{A sparse matrix of class \code{dgCMatrix} or \code{sparseMatrix}.}

\item{directed}{Logical; if \code{TRUE} (default) the graph is directed.}
}
\description{
Create a caugi_graph from a sparse \code{dgCMatrix} or \code{sparseMatrix}
}
