% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_to.R
\name{as_dagitty}
\alias{as_dagitty}
\title{Convert a caugi to a dagitty graph}
\usage{
as_dagitty(x)
}
\arguments{
\item{x}{A \code{caugi} object.}
}
\value{
A \code{dagitty} object.
}
\description{
Convert a caugi to a dagitty graph
}
\examples{
cg <- caugi(
  A \%-->\% B,
  class = "DAG"
)
g_dg <- as_dagitty(cg)

}
\seealso{
Other conversions:
\code{\link[=as_adjacency]{as_adjacency()}},
\code{\link[=as_bnlearn]{as_bnlearn()}},
\code{\link[=as_caugi]{as_caugi()}},
\code{\link[=as_igraph]{as_igraph()}}
}
\concept{conversions}
