% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.caugi_graph}
\alias{as_tibble.caugi_graph}
\title{Convert a \code{caugi_graph} to a tibble.}
\usage{
\method{as_tibble}{caugi_graph}(x, collapse = FALSE, collapse_to = "---", ...)
}
\arguments{
\item{x}{A \code{caugi_graph} object.}

\item{collapse}{Logical. If \code{TRUE}, the edges are collapsed to a symmetric form.}

\item{collapse_to}{Character. The string used to represent collapsed edges.}

\item{...}{Additional arguments (not used).}
}
\description{
Convert a \code{caugi_graph} object to a tibble.
The resulting tibble contains the edges of the graph, with columns for the
source node, target node, and edge type.
}
