% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{build.caugi_graph}
\alias{build.caugi_graph}
\title{Build the graph now}
\usage{
\method{build}{caugi_graph}(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object with \code{cg$built = TRUE}.}
}
\description{
If a \code{caugi_graph} has been modified (nodes or edges added or
removed), it is marked as \emph{not built}, i.e \code{cg$built = FALSE}.
This function builds the graph using the Rust backend and updates the
internal pointer to the graph. If the graph is already built, it is returned.
}
