% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-helpers.R
\name{cargo_test_roclet}
\alias{cargo_test_roclet}
\title{Roclet to run cargo tests}
\usage{
cargo_test_roclet()
}
\description{
A roclet that runs \verb{cargo test} in the Rust source directory
when roxygen is run.
}
