% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi.R
\name{caugi}
\alias{caugi}
\title{Create a \code{caugi} from edge expressions.}
\usage{
caugi(
  ...,
  from = NULL,
  edge = NULL,
  to = NULL,
  nodes = NULL,
  edges_df = NULL,
  simple = TRUE,
  build = NULL,
  class = c("AUTO", "DAG", "UG", "PDAG", "ADMG", "AG", "UNKNOWN"),
  state = NULL,
  .session = NULL
)
}
\arguments{
\item{...}{Edge expressions using the supported infix operators, or
nodes given by symbols or strings. Multiple edges can be
combined using \code{+}: \verb{A --> B + C}, indicating an edge from \code{A} to both \code{B}
and \code{C}. Nodes can also be grouped using \code{c(...)} or parentheses.}

\item{from}{Character vector of source node names.
Optional; mutually exclusive with \code{...}.}

\item{edge}{Character vector of edge types.
Optional; mutually exclusive with \code{...}.}

\item{to}{Character vector of target node names.
Optional; mutually exclusive with \code{...}.}

\item{nodes}{Character vector of node names to declare as isolated nodes.
An optional, but recommended, option is to provide all node names in the
graph, including those that appear in edges. If \code{nodes} is provided, the
order of nodes in the graph will follow the order in \code{nodes}.}

\item{edges_df}{Optional data.frame or data.table with columns
\code{from}, \code{edge}, and \code{to} to specify edges. Mutually exclusive with \code{...}
and \code{from}, \code{edge}, \code{to}. Can be used to create graphs using \code{edges(cg)}
from another \code{caugi} object, \code{cg}.}

\item{simple}{Logical; if \code{TRUE} (default), the graph is a simple graph, and
the function will throw an error if the input contains parallel edges or
self-loops.}

\item{build}{Deprecated.}

\item{class}{Character; one of \code{"AUTO"}, \code{"DAG"}, \code{"UG"}, \code{"PDAG"}, \code{"ADMG"},
\code{"AG"}, or \code{"UNKNOWN"}. \code{"AUTO"} will automatically pick the appropriate
class based on the first match in the order of \code{"DAG"}, \code{"UG"}, \code{"PDAG"},
\code{"ADMG"}, and \code{"AG"}.
It will default to \code{"UNKNOWN"} if no match is found.}

\item{state}{Deprecated. Replaced by \code{.session}.}

\item{.session}{For internal use. Build a graph by supplying a
pre-constructed session pointer from Rust.}
}
\value{
A \code{caugi} S7 object containing the nodes, edges, and a
pointer to the underlying Rust graph structure.
}
\description{
Create a \code{caugi} from a series of edge expressions using
infix operators. Nodes can be specified as symbols, strings, or numbers.

The following edge operators are supported by default:
\itemize{
\item \verb{\%-->\%} for directed edges (A --> B)
\item \verb{\%---\%} for undirected edges (A --- B)
\item \verb{\%<->\%} for bidirected edges (A <-> B)
\item \verb{\%o->\%} for partially directed edges (A o-> B)
\item \verb{\%--o\%} for partially undirected edges (A --o B)
\item \verb{\%o-o\%} for partial edges (A o-o B)
}

You can register additional edge types using \code{\link[=register_caugi_edge]{register_caugi_edge()}}.
}
\examples{
# create a simple DAG (using NSE)
cg <- caugi(
  A \%-->\% B + C,
  B \%-->\% D,
  class = "DAG"
)

# create a PDAG with undirected edges (using NSE)
cg2 <- caugi(
  A \%-->\% B + C,
  B \%---\% D,
  E, # no neighbors for this node
  class = "PDAG"
)

# create a DAG (using SE)
cg3 <- caugi(
  from = c("A", "A", "B"),
  edge = c("-->", "-->", "-->"),
  to = c("B", "C", "D"),
  nodes = c("A", "B", "C", "D", "E"),
  class = "DAG"
)

# create a non-simple graph
cg4 <- caugi(
  A \%-->\% B,
  B \%-->\% A,
  class = "UNKNOWN",
  simple = FALSE
)

cg4@simple # FALSE
cg4@graph_class # "UNKNOWN"

}
\concept{caugi}
