% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{caugi_default_options}
\alias{caugi_default_options}
\title{Default options for caugi}
\usage{
caugi_default_options()
}
\value{
A list of default options for caugi.
}
\description{
Returns the default options for the caugi package. Useful for resetting
options to their original state.
}
\examples{
# Get defaults
caugi_default_options()

# Reset to defaults
caugi_options(caugi_default_options())
}
\seealso{
\code{\link[=caugi_options]{caugi_options()}} for setting and getting options
}
\concept{options}
