% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_from.R
\name{caugi_from_adj_matrix}
\alias{caugi_from_adj_matrix}
\title{Create a caugi_graph from an adjacency matrix}
\usage{
caugi_from_adj_matrix(x, directed = TRUE)
}
\arguments{
\item{x}{A binary adjacency matrix (0/1) or a sparseMatrix}
}
\description{
Create a \code{caugi_graph} from a binary adjacency matrix. Can be
both a \code{sparseMatrix}, \code{dgCMatrix}, or a dense matrix.
}
