% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{caugi_from_dense}
\alias{caugi_from_dense}
\title{Create a caugi_graph from an integer-coded dense matrix}
\usage{
caugi_from_dense(mat)
}
\arguments{
\item{mat}{A dense matrix with integer edge codes or a pcalg::amat}
}
\description{
Create a caugi_graph from an integer-coded dense matrix
}
