% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{caugi_from_graphNEL}
\alias{caugi_from_graphNEL}
\title{Create a caugi_graph from a possibly‐mixed graphNEL}
\usage{
caugi_from_graphNEL(g, collapse = TRUE, collapse_to = "---")
}
\arguments{
\item{g}{A graphNEL object}

\item{collapse}{Logical, whether to collapse directed edges to undirected}

\item{collapse_to}{The code to use for collapsed directed pairs (default "---")}
}
\description{
Create a caugi_graph from a possibly‐mixed graphNEL
}
