% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{caugi_from_sparse}
\alias{caugi_from_sparse}
\title{Create a caugi_graph from a sparse dgCMatrix or sparseMatrix}
\usage{
caugi_from_sparse(mat, directed)
}
\description{
Create a caugi_graph from a sparse dgCMatrix or sparseMatrix
}
