% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{caugi_graph}
\alias{caugi_graph}
\title{Create a caugi_graph (now with \verb{\%-->\%} / \verb{\%<->\%} / \verb{\%---\%} support)}
\usage{
caugi_graph(...)
}
\arguments{
\item{...}{Either the classic \verb{nodes, edges} pair \emph{or} any number of
\code{caugi_edge_spec} objects (built by the new infix operators)
and/or old two-sided formulas (\code{A ~ B + C}).  A leading node
tibble is still allowed.}
}
\value{
A \code{caugi_graph}
}
\description{
Create a caugi_graph (now with \verb{\%-->\%} / \verb{\%<->\%} / \verb{\%---\%} support)
}
