% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{caugi_graph_from_csr}
\alias{caugi_graph_from_csr}
\title{Internal helper: wrap a CSR list and node names into a caugi_graph}
\usage{
caugi_graph_from_csr(node_names, ptrs)
}
\description{
Internal helper: wrap a CSR list and node names into a caugi_graph
}
\keyword{internal}
