% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{check_edge_integer}
\alias{check_edge_integer}
\title{Check if edge type integer maps to edge type}
\usage{
check_edge_integer(x)
}
\description{
Check if edge type integer maps to edge type
}
\keyword{internal}
