% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_directed_PAG_to_symmetric_PAG}
\alias{collapse_directed_PAG_to_symmetric_PAG}
\title{Collapse mutual directed PAG edges (o->) to a single o-o edge}
\usage{
collapse_directed_PAG_to_symmetric_PAG(df)
}
\arguments{
\item{df}{Tibble with columns \code{from}, \code{to}, \code{edge_type}}
}
\value{
Tibble with collapsed edge list
}
\description{
Collapse mutual directed PAG edges (o->) to a single o-o edge
}
\keyword{internal}
