% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_directed_to_symmetric_edges}
\alias{collapse_directed_to_symmetric_edges}
\title{Collapse mutual directed edges (-->, not o->) to a single edge (--- or <->)}
\usage{
collapse_directed_to_symmetric_edges(df, collapse_to = "---")
}
\arguments{
\item{df}{A tibble with columns from, to, edge_type}

\item{collapse_to}{The code to use for collapsed directed pairs (default "---")}
}
\value{
A tibble with directed pairs collapsed
}
\description{
Collapse mutual directed edges (-->, not o->) to a single edge (--- or <->)
}
\keyword{internal}
