% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{collapse_symmetric_edges}
\alias{collapse_symmetric_edges}
\title{Collapse bidirected (<->) and undirected (---) edges to single edges}
\usage{
collapse_symmetric_edges(df)
}
\arguments{
\item{df}{A tibble with columns from, to, edge_type}
}
\value{
A tibble with symmetric edges collapsed
}
\description{
Collapse bidirected (<->) and undirected (---) edges to single edges
}
\keyword{internal}
