% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{convert_graph_class}
\alias{convert_graph_class}
\title{Convert a \code{caugi_graph} to a specified graph type, if possible.}
\usage{
convert_graph_class(cg, class)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{class}{The target graph class. One of \code{"DAG"}, \code{"PDAG"}, \code{"Unknown"}.}
}
\value{
The converted \code{caugi_graph} object.
}
\description{
Convert a \code{caugi_graph} to a specified graph type, if possible. It throws an
error if the conversion is not possible. For example, a graph with undirected
edges cannot be converted to a DAG.
}
