% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaaa-coding-helpers.R
\name{coverage_report}
\alias{coverage_report}
\title{Combined Coverage Report for R and Rust Code}
\usage{
coverage_report(
  pkg = ".",
  rust_dir = "src/rust",
  run_r = TRUE,
  run_rust = TRUE,
  ignore = character(),
  ...
)
}
\arguments{
\item{pkg}{Character. The path to the R package.
Default is the current directory \code{"."}.}

\item{rust_dir}{Character. The path to the Rust source directory within the
package.}

\item{run_r}{Logical. Whether to run coverage for R code. Defaults to TRUE.}

\item{run_rust}{Logical. Whether to run coverage for Rust code.
Defaults to TRUE.}

\item{ignore}{Character vector. File patterns to ignore in the coverage
report. For example, ignoring \code{lib.rs} gives a more meaningful coverage
estimate. Defaults to \code{character()}.}

\item{...}{Additional arguments passed to \code{covr::report()}.}
}
\value{
Invisibly returns a \code{covr} coverage object.
}
\description{
Generates a combined code coverage report for R and Rust code in a package
using \code{covr}. This function runs coverage analysis for both R and Rust code,
merges the results, and produces a unified coverage report. Currently, it
does not support coverage from R tests of Rust code.
}
