% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaaa-coding-helpers.R
\name{document_caugi}
\alias{document_caugi}
\title{Run devtools::document() with cargo tests roclet}
\usage{
document_caugi()
}
\description{
Run devtools::document() with cargo tests roclet
}
\keyword{internal}
