% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.capture_nodes_expr}
\alias{.capture_nodes_expr}
\title{Capture nodes expression}
\usage{
.capture_nodes_expr(expr, env = parent.frame())
}
\arguments{
\item{expr}{An expression representing the target node(s).}

\item{env}{The environment in which to evaluate the expression.}
}
\value{
A character vector of target node names.
}
\description{
Capture and evaluate an expression representing nodes.
}
\keyword{internal}
