% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.caugi_builtin_specs}
\alias{.caugi_builtin_specs}
\title{Built-in edge specifications}
\usage{
.caugi_builtin_specs()
}
\value{
A tibble with columns \code{glyph}, \code{left_mark}, \code{right_mark},
\code{orientation}, \code{class}, \code{symmetric}, and \code{traversable_when_conditioned}.
}
\description{
A tibble of the built-in edge specifications.
}
\keyword{internal}
