% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.caugi_fit_on_line}
\alias{.caugi_fit_on_line}
\title{Fit items on a line}
\usage{
.caugi_fit_on_line(items, width, indent)
}
\arguments{
\item{items}{A character vector of item labels.}

\item{width}{An integer specifying the total line width.}

\item{indent}{An integer specifying the indent width.}
}
\value{
An integer indicating how many items fit on the line.
}
\description{
Helper function to determine how many items can fit on a line
of given width, considering an indent.
}
\keyword{internal}
