% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.collect_edges_nodes}
\alias{.collect_edges_nodes}
\title{Collect edges and nodes}
\usage{
.collect_edges_nodes(calls)
}
\arguments{
\item{calls}{A list of expressions from caugi(...)}
}
\value{
A list with two elements:
\itemize{
\item edges: a tibble with columns \code{from}, \code{edge}, \code{to}
\item declared: a character vector of explicitly declared nodes
}
}
\description{
Collect edges (via .parse_edge_arg) and explicitly declared
nodes (no edges).
}
\keyword{internal}
