% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.combine_plus}
\alias{.combine_plus}
\title{Combine terms with '+'}
\usage{
.combine_plus(terms)
}
\arguments{
\item{terms}{A list of expressions to combine.}
}
\value{
A single expression combining the terms with '+'.
}
\description{
Combine a list of terms into a single left-associative '+' call.
}
\keyword{internal}
