% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.contains_edge}
\alias{.contains_edge}
\title{Does the expression contain an edge call?}
\usage{
.contains_edge(expr)
}
\arguments{
\item{expr}{An expression to check}
}
\value{
TRUE if the expression contains an edge call, FALSE otherwise
}
\description{
Recursively check if the expression contains any edge call.
}
\keyword{internal}
