% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.edge_constructor}
\alias{.edge_constructor}
\title{Edge constructor}
\usage{
.edge_constructor(from = character(), edge = character(), to = character())
}
\arguments{
\item{from}{Character vector of source node names.}

\item{edge}{Character vector of edge glyphs.}

\item{to}{Character vector of target node names.}
}
\value{
A \code{data.table} object with columns \code{from}, \code{edge}, and \code{to}.
}
\description{
Internal function to construct edges for \code{caugi} objects.
}
\keyword{internal}
