% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.edge_constructor_idx}
\alias{.edge_constructor_idx}
\title{Edge constructor using indices.}
\usage{
.edge_constructor_idx(from_idx, edge, to_idx, node_names)
}
\arguments{
\item{from_idx}{Integer vector of source node indices.}

\item{edge}{Character vector of edge glyphs.}

\item{to_idx}{Integer vector of target node indices.}
}
\value{
A \code{data.table} object with columns \code{from}, \code{edge}, and \code{to}.
}
\description{
Internal function to construct edges for \code{caugi} objects using indices.
}
