% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-env.R
\name{.edge_ops_get}
\alias{.edge_ops_get}
\title{Get edge operators}
\usage{
.edge_ops_get()
}
\value{
The current edge operators of the caugi environment
}
\description{
This function gets the default caugi edge operators
}
\keyword{internal}
