% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.edge_units_to_dt}
\alias{.edge_units_to_dt}
\title{Turn edge units into a \code{data.table} of edges}
\usage{
.edge_units_to_dt(units)
}
\arguments{
\item{units}{A list of edge units, each with \code{lhs}, \code{rhs}, and \code{glyph}.}
}
\value{
A \code{data.table} with columns \code{from}, \code{edge}, and \code{to}.
}
\description{
Convert a list of edge units into a \code{data.table} with columns
\code{from}, \code{edge}, and \code{to}.
}
\keyword{internal}
