% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.expand_nodes}
\alias{.expand_nodes}
\title{Expand node expressions}
\usage{
.expand_nodes(expr)
}
\arguments{
\item{expr}{An expression representing nodes.}
}
\value{
A character vector of node names.
}
\description{
Expand node expressions: symbol, "B", c(...), +, (...)
}
\keyword{internal}
