% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.flatten_edge_chain}
\alias{.flatten_edge_chain}
\title{Flatten a chained edge expression}
\usage{
.flatten_edge_chain(call_expr)
}
\arguments{
\item{call_expr}{A call expression representing a chained edge.}
}
\value{
A list with two elements, \code{terms} and \code{ops}.
}
\description{
Given a chained edge expression,
flatten it into its terms and operators.
}
\keyword{internal}
