% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{.freeze_state}
\alias{.freeze_state}
\title{Internal: Freeze the state environment of a \code{caugi_graph}}
\usage{
.freeze_state(e)
}
\arguments{
\item{e}{The state environment to freeze.}
}
\value{
The frozen environment.
}
\description{
Internal function to freeze the state environment of a
\code{caugi_graph}, preventing further modifications. This function is not
intended to be used directly by users.
}
\keyword{internal}
