% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{.get_edges}
\alias{.get_edges}
\title{Build edges \code{data.table} from verb call.}
\usage{
.get_edges(from, edge, to, calls, simple = TRUE)
}
\arguments{
\item{from}{Character vector of source node names.}

\item{edge}{Character vector of edge types.}

\item{to}{Character vector of target node names.}

\item{calls}{List of calls from \code{...}.}

\item{simple}{Logical, whether the graph is simple or not.}
}
\value{
A \code{data.table} with columns \code{from}, \code{edge}, and \code{to}.
}
\description{
Internal helper to build edges \code{data.table} from verb call.
}
\keyword{internal}
