% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{.get_nodes}
\alias{.get_nodes}
\title{Get nodes \code{data.table} from verb call.}
\usage{
.get_nodes(name, calls)
}
\arguments{
\item{name}{Character vector of node names.}

\item{calls}{List of calls from \code{...}.}
}
\value{
A \code{data.table} with column \code{name} for node names.
}
\description{
Internal helper to build nodes \code{data.table} from verb call.
}
\keyword{internal}
