% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{.get_nodes_tibble}
\alias{.get_nodes_tibble}
\title{Get nodes tibble from verb call.}
\usage{
.get_nodes_tibble(name, calls)
}
\arguments{
\item{name}{Character vector of node names.}

\item{calls}{List of calls from \code{...}.}
}
\value{
A tibble with column \code{name} for node names.
}
\description{
Internal helper to build nodes tibble from verb call.
}
\keyword{internal}
