% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.getter_output}
\alias{.getter_output}
\title{Output object of getter queries}
\usage{
.getter_output(cg, idx0)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{idx0}{A vector of zero-based node indices.}
}
\value{
A tibble with a \code{name} column.
}
\description{
Helper to format the output of getter queries.
}
\keyword{internal}
