% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-env.R
\name{.glyph_map_get}
\alias{.glyph_map_get}
\title{Get edge operators}
\usage{
.glyph_map_get()
}
\value{
The current edge glyphs of the caugi environment
}
\description{
This function gets the default caugi edge glyphs
}
\keyword{internal}
