% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.glyph_of}
\alias{.glyph_of}
\title{Get glyph for an operator}
\usage{
.glyph_of(op_sym)
}
\arguments{
\item{op_sym}{A symbol representing the edge operator (e.g., \verb{\%-->\%}).}
}
\value{
A string representing the edge glyph (e.g., \code{"-->"}).
}
\description{
Get the glyph string for a given edge operator symbol.
}
\keyword{internal}
