% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.is_edge_call}
\alias{.is_edge_call}
\title{Is it an edge call / expression?}
\usage{
.is_edge_call(expr)
}
\arguments{
\item{expr}{An expression to check}
}
\value{
TRUE if the expression is an edge call, FALSE otherwise
}
\description{
This function checks if the expression is an edge call
}
\keyword{internal}
