% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.is_node_expr}
\alias{.is_node_expr}
\title{Is it a node expr?}
\usage{
.is_node_expr(expr)
}
\arguments{
\item{expr}{An expression to check}
}
\value{
TRUE if the expression is a valid node expression, FALSE otherwise
}
\description{
Check if the expression is a valid node expression:
symbol, string, number, c(...), +, (...)
}
\keyword{internal}
