% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{.mark_not_built}
\alias{.mark_not_built}
\title{Mark a \code{caugi_graph} as \emph{not built}.}
\usage{
.mark_not_built(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
The same \code{caugi_graph} object, but with the \code{built} attribute set to
\code{FALSE}.
}
\description{
When a \code{caugi_graph} is modified, it should be marked as not
built. This function sets the \code{built} attribute to \code{FALSE}. Thereby, the Rust
backend and the R frontend does not match, and at one point, the
\code{caugi_graph} will need to be rebuild for it to be queried.
}
\keyword{internal}
