% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.node_constructor}
\alias{.node_constructor}
\title{Node constructor}
\usage{
.node_constructor(names = character(), sort = FALSE)
}
\arguments{
\item{names}{Character vector of node names.}

\item{sort}{Logical indicating whether to sort the node names.}
}
\value{
A \code{data.table} object with a single column \code{name}.
}
\description{
A simple wrapper creating a \code{data.table} object with a single column \code{name}.
}
\details{
The reason this exists is so if changes should be made in the future, it is
easy to simply change this constructor, rather than changing the calls to
\code{data.table} all over the place.
}
\keyword{internal}
