% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.relations}
\alias{.relations}
\title{Get relations of nodes in a \code{caugi_graph}}
\usage{
.relations(cg, nodes_expr, index, getter)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{nodes_expr}{A node expression from .expand_nodes.}

\item{index}{A vector of node indices.}

\item{getter}{A function that takes a pointer to the graph and a zero-based
node index, and returns a vector of zero-based indices of related nodes.}
}
\value{
A list of character vectors, each a set of related node names.
}
\description{
Helper to get relations (parents, children, etc.) of nodes.
}
\keyword{internal}
