% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{.resolve_idx0_get}
\alias{.resolve_idx0_get}
\alias{.resolve_idx0_mget}
\title{Resolve node name or index to 0-based index.}
\usage{
.resolve_idx0_get(nm_idx_map, node_name = NULL, node_index = NULL)

.resolve_idx0_mget(nm_idx_map, node_name = NULL, node_index = NULL)
}
\arguments{
\item{nm_idx_map}{A \code{fastmap} mapping node names to 0-based indices from
a \code{caugi}.}

\item{node_name}{Optional character vector of node names.}

\item{node_index}{Optional numeric vector of 1-based node indices.}
}
\description{
Internal helper function to resolve either a node name or a
node index to a 0-based index.
\code{.resolve_idx0_get} uses \code{get} on the \code{fastmap} and expects a single value,
while \code{.resolve_idx0_mget} uses \code{mget} and can return multiple values.
}
\seealso{
\link[fastmap:fastmap]{fastmap::fastmap}
}
\keyword{internal}
