% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.resolve_idx_from_index}
\alias{.resolve_idx_from_index}
\title{Resolve 1-based indices to zero-based}
\usage{
.resolve_idx_from_index(cg, index)
}
\description{
Resolve 1-based indices to zero-based
}
\keyword{internal}
