% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.segment_units}
\alias{.segment_units}
\title{Create an edge unit from lhs, op, rhs}
\usage{
.segment_units(lhs_term, op_chr, rhs_term)
}
\arguments{
\item{lhs_term}{An expression for the left-hand side nodes.}

\item{op_chr}{A string representing the edge operator glyph.}

\item{rhs_term}{An expression for the right-hand side nodes.}
}
\value{
A list with elements \code{lhs}, \code{rhs}, and \code{glyph}.
}
\description{
Create an edge unit from lhs, op, rhs expressions.
}
\keyword{internal}
