% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi.R
\name{.session_to_caugi}
\alias{.session_to_caugi}
\title{Convert a graph session to a \code{caugi} S7 object}
\usage{
.session_to_caugi(session, node_names = NULL)
}
\arguments{
\item{session}{A pointer to the underlying Rust GraphSession.}

\item{node_names}{Optional character vector of node names. If \code{NULL}
(default), node names will be taken from the session.}
}
\value{
A \code{caugi} object representing the graph.
}
\description{
Convert a graph pointer from Rust to a \code{caugi} to a
S7 object.
}
\keyword{internal}
