% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.set_names}
\alias{.set_names}
\title{Set names to an object}
\usage{
.set_names(object = nm, nm)
}
\arguments{
\item{object}{An R object to which names are to be assigned.}

\item{nm}{A character vector of names to assign to the object.}
}
\value{
The input object with the assigned names.
}
\description{
Only made to avoid using \code{stats::setNames}, so that the \code{stats} package is
not a dependency.
}
\keyword{internal}
