% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints-solver.R
\name{.solver_supported_classes}
\alias{.solver_supported_classes}
\title{Supported graph classes for the constraint solver.}
\usage{
.solver_supported_classes()
}
\description{
Supported graph classes for the constraint solver.
}
