% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL-parser.R
\name{.split_plus}
\alias{.split_plus}
\title{Expand target expressions with \code{=}}
\usage{
.split_plus(expr)
}
\arguments{
\item{expr}{An expression representing nodes.}
}
\value{
A character vector of node names.
}
\description{
Split any expression into top-level '+' terms (fully flattened).
}
\keyword{internal}
