% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{.view_to_caugi_graph}
\alias{.view_to_caugi_graph}
\title{Convert a graph pointer to a \code{caugi_graph} S7 object}
\usage{
.view_to_caugi_graph(ptr, node_names = NULL)
}
\arguments{
\item{ptr}{A pointer to the underlying Rust graph structure.}

\item{node_names}{Optional character vector of node names. If \code{NULL}
(default), nodes will be named \code{V1}, \code{V2}, ..., \code{Vn}.}
}
\value{
A \code{caugi_graph} object representing the graph.
}
\description{
Convert a graph pointer from Rust to a \code{caugi_graph} to a
S7 object.
}
\keyword{internal}
