% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_operators.R
\name{edge_operators}
\alias{edge_operators}
\alias{\%-->\%}
\alias{\%---\%}
\alias{\%<->\%}
\alias{\%o-o\%}
\alias{\%--o\%}
\alias{\%o->\%}
\title{Infix operators for edge specifications}
\usage{
lhs \%-->\% rhs

lhs \%---\% rhs

lhs \%<->\% rhs

lhs \%o-o\% rhs

lhs \%--o\% rhs

lhs \%o->\% rhs
}
\arguments{
\item{lhs}{The left-hand side node expression.}

\item{rhs}{The right-hand side node expression.}
}
\value{
A \code{data.table} with columns \code{from}, \code{to}, and \code{edge}.
}
\description{
These operators are used to specify edges in \code{caugi()}.
Should be used internally in \code{caugi()} calls.
}
\keyword{internal}
