% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_types.R
\docType{data}
\name{edge_type_levels}
\alias{edge_type_levels}
\title{Edge type levels}
\format{
An object of class \code{character} of length 6.
}
\usage{
edge_type_levels
}
\description{
Edge type levels
}
\keyword{internal}
