% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi.R
\name{freeze_state}
\alias{freeze_state}
\alias{.freeze_state}
\alias{.unfreeze_state}
\title{Internal: Freeze the state environment of a \code{caugi}}
\usage{
.freeze_state(e)

.unfreeze_state(e)
}
\arguments{
\item{e}{The state environment to freeze/unfreeze.}
}
\value{
The frozen/unfrozen environment.
}
\description{
Internal functions to freeze and unfreeze the state environment
of a \code{caugi}, preventing further modifications. These functions are not
intended to be used directly by users.
}
\keyword{internal}
