% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_operators.R
\name{\%-->\%}
\alias{\%-->\%}
\title{Infix operators for edge specifications}
\usage{
lhs \%-->\% rhs
}
\arguments{
\item{lhs}{The left-hand side node expression.}

\item{rhs}{The right-hand side node expression.}
}
\value{
A tibble with columns \code{from}, \code{to}, and \code{edge}.
}
\description{
These operators are used to specify edges in \code{caugi_graph()}.
Should be used internally in \code{caugi_graph()} calls.
}
\keyword{internal}
