% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_caugi}
\alias{is_caugi}
\title{Is it a \code{caugi} graph?}
\usage{
is_caugi(x, throw_error = FALSE)
}
\arguments{
\item{x}{An object to check.}

\item{throw_error}{Logical; if \code{TRUE}, throws an error if \code{x} is not a
\code{caugi}.}
}
\value{
A logical value indicating whether the object is a \code{caugi}.
}
\description{
Checks if the given object is a \code{caugi}. Mostly used
internally to validate inputs.
}
\examples{
cg <- caugi(
  A \%-->\% B,
  class = "DAG"
)

is_caugi(cg) # TRUE

}
\seealso{
Other queries:
\code{\link[=ancestors]{ancestors()}},
\code{\link[=anteriors]{anteriors()}},
\code{\link[=children]{children()}},
\code{\link[=descendants]{descendants()}},
\code{\link[=districts]{districts()}},
\code{\link[=edge_types]{edge_types()}},
\code{\link[=edges]{edges()}},
\code{\link[=exogenous]{exogenous()}},
\code{\link[=is_acyclic]{is_acyclic()}},
\code{\link[=is_admg]{is_admg()}},
\code{\link[=is_ag]{is_ag()}},
\code{\link[=is_cpdag]{is_cpdag()}},
\code{\link[=is_dag]{is_dag()}},
\code{\link[=is_empty_caugi]{is_empty_caugi()}},
\code{\link[=is_mag]{is_mag()}},
\code{\link[=is_mpdag]{is_mpdag()}},
\code{\link[=is_pdag]{is_pdag()}},
\code{\link[=is_simple]{is_simple()}},
\code{\link[=is_ug]{is_ug()}},
\code{\link[=m_separated]{m_separated()}},
\code{\link[=markov_blanket]{markov_blanket()}},
\code{\link[=neighbors]{neighbors()}},
\code{\link[=nodes]{nodes()}},
\code{\link[=parents]{parents()}},
\code{\link[=posteriors]{posteriors()}},
\code{\link[=same_nodes]{same_nodes()}},
\code{\link[=spouses]{spouses()}},
\code{\link[=subgraph]{subgraph()}},
\code{\link[=topological_sort]{topological_sort()}}
}
\concept{queries}
