% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_cpdag}
\alias{is_cpdag}
\title{Is the \code{caugi} graph a CPDAG?}
\usage{
is_cpdag(cg)
}
\arguments{
\item{cg}{A \code{caugi} object.}
}
\value{
A logical value indicating whether the graph is a CPDAG.
}
\description{
Checks if the given \code{caugi} graph is a
Complete Partially Directed Acyclic Graph (CPDAG).
}
\examples{
cg_cpdag <- caugi(
  A \%---\% B,
  A \%-->\% C,
  B \%-->\% C,
  class = "PDAG"
)
is_cpdag(cg_cpdag) # TRUE

cg_not_cpdag <- caugi(
  A \%---\% B,
  A \%---\% C,
  B \%-->\% C,
  class = "PDAG"
)
is_cpdag(cg_not_cpdag) # FALSE

}
\references{
C. Meek (1995). Causal inference and causal explanation with background
knowledge. In \emph{Proceedings of the Eleventh Conference on Uncertainty in
Artificial Intelligence (UAI-95)}, pp. 403--411. Morgan Kaufmann.
}
\seealso{
Other queries:
\code{\link[=ancestors]{ancestors()}},
\code{\link[=anteriors]{anteriors()}},
\code{\link[=children]{children()}},
\code{\link[=descendants]{descendants()}},
\code{\link[=districts]{districts()}},
\code{\link[=edge_types]{edge_types()}},
\code{\link[=edges]{edges()}},
\code{\link[=exogenous]{exogenous()}},
\code{\link[=is_acyclic]{is_acyclic()}},
\code{\link[=is_admg]{is_admg()}},
\code{\link[=is_ag]{is_ag()}},
\code{\link[=is_caugi]{is_caugi()}},
\code{\link[=is_dag]{is_dag()}},
\code{\link[=is_empty_caugi]{is_empty_caugi()}},
\code{\link[=is_mag]{is_mag()}},
\code{\link[=is_mpdag]{is_mpdag()}},
\code{\link[=is_pdag]{is_pdag()}},
\code{\link[=is_simple]{is_simple()}},
\code{\link[=is_ug]{is_ug()}},
\code{\link[=m_separated]{m_separated()}},
\code{\link[=markov_blanket]{markov_blanket()}},
\code{\link[=neighbors]{neighbors()}},
\code{\link[=nodes]{nodes()}},
\code{\link[=parents]{parents()}},
\code{\link[=posteriors]{posteriors()}},
\code{\link[=same_nodes]{same_nodes()}},
\code{\link[=spouses]{spouses()}},
\code{\link[=subgraph]{subgraph()}},
\code{\link[=topological_sort]{topological_sort()}}
}
\concept{queries}
