% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{is_edge_symmetric}
\alias{is_edge_symmetric}
\title{Is the edge symmetric?}
\usage{
is_edge_symmetric(glyph)
}
\arguments{
\item{glyph}{A string representing the edge glyph (e.g., \code{"-->"}, \code{"<->"}).}
}
\value{
Logical, \code{TRUE} if the edge is symmetric, otherwise throws error.
}
\description{
Check if the given edge glyph is symmetric in the edge registry.
}
\keyword{internal}
