% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{length.caugi_graph}
\alias{length.caugi_graph}
\title{Length of a \code{caugi_graph}}
\usage{
\method{length}{caugi_graph}(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
An integer representing the number of nodes.
}
\description{
Returns the number of nodes in the graph.
}
