% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{moralize}
\alias{moralize}
\title{Moralize a DAG}
\usage{
moralize(cg)
}
\arguments{
\item{cg}{A \code{caugi} object (DAG).}
}
\value{
A \code{caugi} object representing the moralized graph (UG).
}
\description{
Moralizing a DAG involves connecting all parents of each node and then
converting all directed edges into undirected edges.
}
\details{
This changes the graph from a Directed Acyclic Graph (DAG) to an
Undirected Graph (UG), also known as a Markov Graph.
}
\examples{
cg <- caugi(A \%-->\% C, B \%-->\% C, class = "DAG")
moralize(cg) # A -- B, A -- C, B -- C

}
\seealso{
Other operations:
\code{\link[=condition_marginalize]{condition_marginalize()}},
\code{\link[=count_dags]{count_dags()}},
\code{\link[=dag_from_pdag]{dag_from_pdag()}},
\code{\link[=enumerate_dags]{enumerate_dags()}},
\code{\link[=exogenize]{exogenize()}},
\code{\link[=latent_project]{latent_project()}},
\code{\link[=meek_closure]{meek_closure()}},
\code{\link[=mutate_caugi]{mutate_caugi()}},
\code{\link[=normalize_latent_structure]{normalize_latent_structure()}},
\code{\link[=skeleton]{skeleton()}}
}
\concept{operations}
