% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{mutate_caugi}
\alias{mutate_caugi}
\title{Mutate \code{caugi} class}
\usage{
mutate_caugi(cg, class)
}
\arguments{
\item{cg}{A \code{caugi} object.}

\item{class}{A character string specifying the new class.}
}
\value{
A \code{caugi} object of the specified class.
}
\description{
Mutate the \code{caugi} class from one graph class to another, if possible.
For example, convert a \code{DAG} to a \code{PDAG}, or a fully directed \code{caugi} of
class \code{UNKNOWN} to a \code{DAG}. Throws an error if not possible.
}
\details{
This function returns a copy of the object, and the original remains
unchanged.
}
\examples{
cg <- caugi(A \%-->\% B, class = "UNKNOWN")
cg_dag <- mutate_caugi(cg, "DAG")

}
\seealso{
Other operations:
\code{\link[=condition_marginalize]{condition_marginalize()}},
\code{\link[=count_dags]{count_dags()}},
\code{\link[=dag_from_pdag]{dag_from_pdag()}},
\code{\link[=enumerate_dags]{enumerate_dags()}},
\code{\link[=exogenize]{exogenize()}},
\code{\link[=latent_project]{latent_project()}},
\code{\link[=meek_closure]{meek_closure()}},
\code{\link[=moralize]{moralize()}},
\code{\link[=normalize_latent_structure]{normalize_latent_structure()}},
\code{\link[=skeleton]{skeleton()}}
}
\concept{operations}
