% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{nodes}
\alias{nodes}
\alias{vertices}
\alias{V}
\alias{edges}
\alias{E}
\title{Get nodes or edges of a \code{caugi_graph}}
\usage{
nodes(cg)

vertices(cg)

V(cg)

edges(cg)

E(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
\strong{nodes}: A tibble with a \code{name} column.
\strong{edges}: A tibble with \code{from}, \code{edge}, and \code{to} columns.
}
\description{
Get nodes or edges of a \code{caugi_graph}.
}
