% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{reset_caugi_registry}
\alias{reset_caugi_registry}
\title{Reset the global edge registry.}
\usage{
reset_caugi_registry()
}
\value{
TRUE, invisibly.
}
\description{
This is mainly useful for testing purposes to clear the edge
registry.
}
