% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-helpers.R
\name{roclet_clean.roclet_cargo_test}
\alias{roclet_clean.roclet_cargo_test}
\title{Clean for the cargo tests roclet}
\usage{
roclet_clean.roclet_cargo_test(x, base_path)
}
\description{
This function is a no-op because there is nothing to clean up.
}
