% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-helpers.R
\name{roclet_output.roclet_cargo_test}
\alias{roclet_output.roclet_cargo_test}
\title{Output for the cargo tests roclet}
\usage{
roclet_output.roclet_cargo_test(x, results, base_path, ...)
}
\description{
This function runs \verb{cargo test} in the Rust source directory if a
\code{Cargo.toml} file is found there.
}
