% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_constraints_consistent}
\alias{rs_constraints_consistent}
\title{Check satisfiability of a list of constraint formulas against a
node set under the given graph-class invariants.}
\usage{
rs_constraints_consistent(formulas, nodes, class)
}
\description{
Check satisfiability of a list of constraint formulas against a
node set under the given graph-class invariants.
}
