% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_constraints_enumerate}
\alias{rs_constraints_enumerate}
\title{Enumerate up to \code{limit} graphs of the given class satisfying the
constraints. Returns a list of edge data frames.}
\usage{
rs_constraints_enumerate(formulas, nodes, class, limit)
}
\description{
Enumerate up to \code{limit} graphs of the given class satisfying the
constraints. Returns a list of edge data frames.
}
