% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaaa-coding-helpers.R
\name{run_cargo_test}
\alias{run_cargo_test}
\title{Run \verb{cargo test} in the Rust source directory manually.}
\usage{
run_cargo_test(path = "./src/rust")
}
\arguments{
\item{path}{Character. The path to the Rust source directory.
Default is "./src/rust".}
}
\value{
The output of processx::run command
}
\description{
This function runs \verb{cargo test} in the specified Rust source
directory using processx.
}
\keyword{internal}
