% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaaa-coding-helpers.R
\name{rust_r_covr}
\alias{rust_r_covr}
\title{Generate Coverage Reports for R and Rust Code}
\usage{
rust_r_covr(pkg = ".", rust_dir = "src/rust", run_r = TRUE, run_rust = TRUE)
}
\arguments{
\item{pkg}{Character. The path to the R package. Default is the current
directory \code{"."}.}

\item{rust_dir}{Character. The path to the Rust source directory within the
package. Default is \code{"src/rust"}.}

\item{run_r}{Logical. Whether to run coverage for R code. Defaults to TRUE.}

\item{run_rust}{Logical. Whether to run coverage for Rust code.
Defaults to TRUE.}
}
\value{
A list with elements \code{r} and \code{rust} containing Cobertura XML strings
for R and Rust coverage, respectively. If only one type of coverage is run,
it returns just the Cobertura XML string for that type.
}
\description{
Generates coverage reports for both R and Rust code in a package using \code{covr}
and \verb{cargo llvm-cov}. It runs coverage analysis for R code using \code{covr}
and for Rust code using \verb{cargo llvm-cov}, then combines the results into a
list containing Cobertura XML strings for both R and Rust coverage.
}
\keyword{internal}
