% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sample_gnm}
\alias{sample_gnm}
\title{Random caugi graph with exactly m edges (G(n, m)) for DAG or CPDAG}
\usage{
sample_gnm(n, m, mode = c("dag", "cpdag"), seed = NULL)
}
\arguments{
\item{n}{Integer >= 0}

\item{m}{Integer in \verb{0, n*(n-1)/2}}

\item{mode}{"dag" or "cpdag"}

\item{seed}{Optional integer seed for reproducibility}
}
\value{
An external pointer to a \code{GraphView} (DAG or PDAG)
}
\description{
Random caugi graph with exactly m edges (G(n, m)) for DAG or CPDAG
}
