% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{seal_caugi_registry}
\alias{seal_caugi_registry}
\title{Seal the global edge registry.}
\usage{
seal_caugi_registry()
}
\description{
Seal the global edge registry to prevent further modifications. Can be used
to ensure reproducability.
}
\seealso{
Other registry: 
\code{\link{caugi_registry}()},
\code{\link{register_caugi_edge}()},
\code{\link{reset_caugi_registry}()}
}
\concept{registry}
