% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{skeleton}
\alias{skeleton}
\title{Get the skeleton of a graph}
\usage{
skeleton(cg)
}
\arguments{
\item{cg}{A \code{caugi} object. Either a DAG or PDAG.}
}
\value{
A \code{caugi} object representing the skeleton of the graph (UG).
}
\description{
The skeleton of a graph is obtained by replacing all directed edges with
undirected edges.
}
\details{
This changes the graph from any class to an Undirected Graph (UG), also known
as a Markov Graph.
}
\examples{
cg <- caugi(A \%-->\% B, class = "DAG")
skeleton(cg) # A --- B

}
\seealso{
Other operations:
\code{\link[=condition_marginalize]{condition_marginalize()}},
\code{\link[=count_dags]{count_dags()}},
\code{\link[=dag_from_pdag]{dag_from_pdag()}},
\code{\link[=enumerate_dags]{enumerate_dags()}},
\code{\link[=exogenize]{exogenize()}},
\code{\link[=latent_project]{latent_project()}},
\code{\link[=meek_closure]{meek_closure()}},
\code{\link[=moralize]{moralize()}},
\code{\link[=mutate_caugi]{mutate_caugi()}},
\code{\link[=normalize_latent_structure]{normalize_latent_structure()}}
}
\concept{operations}
