% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{union.caugi_graph}
\alias{union.caugi_graph}
\title{Return union of two \code{caugi_graphs}}
\usage{
union.caugi_graph(cg1, cg2)
}
\arguments{
\item{cg1}{A \code{caugi_graph} object.}

\item{cg2}{A \code{caugi_graph} object.}
}
\value{
A new \code{caugi_graph} object representing the union of \code{cg1} and
\code{cg2}.
}
\description{
Returns a new \code{caugi_graph} that is the union of two input
graphs.
}
