% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{aid}
\alias{aid}
\title{Adjustment Identification Distance}
\usage{
aid(cg1, cg2, type = c("oset", "ancestor", "parent"))
}
\arguments{
\item{cg1}{A \code{caugi_graph} object.}

\item{cg2}{A \code{caugi_graph} object.}
}
\value{
The AID between the two graphs.
}
\description{
Compute the Adjustment Identification Distance (AID) between two
graphs using the \code{gadjid} Rust package.
}
