% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{aid}
\alias{aid}
\title{Adjustment Identification Distance}
\usage{
aid(truth, guess, type = c("oset", "ancestor", "parent"), normalized = TRUE)
}
\arguments{
\item{truth}{A \code{caugi_graph} object.}

\item{guess}{A \code{caugi_graph} object.}

\item{type}{A character string specifying the type of AID to compute.
Options are \code{"oset"} (default), \code{"ancestor"}, and \code{"parent"}.}

\item{normalized}{Logical; if \code{TRUE}, returns the normalized AID. If \code{FALSE},
returns the count.}
}
\value{
A list containing the score (normalized) and the count.
}
\description{
Compute the Adjustment Identification Distance (AID) between two
graphs using the \code{gadjid} Rust package.
}
\seealso{
Other metrics: 
\code{\link{hd}()},
\code{\link{shd}()}
}
\concept{metrics}
