% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_to.R
\name{as_bnlearn}
\alias{as_bnlearn}
\title{Convert a caugi to a bnlearn network}
\usage{
as_bnlearn(x)
}
\arguments{
\item{x}{A \code{caugi} object.}
}
\value{
A \code{bnlearn} DAG.
}
\description{
Convert a caugi to a bnlearn network
}
\examples{
cg <- caugi(
  A \%-->\% B,
  class = "DAG"
)
g_bn <- as_bnlearn(cg)

}
\seealso{
Other conversions: 
\code{\link{as_adjacency}()},
\code{\link{as_caugi}()},
\code{\link{as_dagitty}()},
\code{\link{as_igraph}()}
}
\concept{conversions}
