% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R, R/caugi_from.R
\name{as_caugi.graphNEL}
\alias{as_caugi.graphNEL}
\alias{caugi_from_graphNEL}
\title{Create a caugi_graph from a possibly‐mixed graphNEL}
\usage{
\method{as_caugi}{graphNEL}(x, collapse = FALSE, collapse_to = "---", ...)

caugi_from_graphNEL(x, collapse = FALSE, collapse_to = "---")
}
\arguments{
\item{x}{A graphNEL object}

\item{collapse}{Logical, whether to collapse directed edges to undirected}

\item{collapse_to}{The code to use for collapsed directed pairs (default "---")}
}
\description{
Create a caugi_graph from a possibly‐mixed graphNEL
}
