% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi}
\alias{as_caugi}
\title{Convert to a \code{caugi_graph}}
\usage{
as_caugi(x, ...)
}
\arguments{
\item{x}{An object to convert to a \code{caugi_graph}.}

\item{...}{Additional arguments passed to specific methods.}
}
\description{
Convert an object to a \code{caugi_graph}. The object can be a
\code{graphNEL}, \code{sparseMatrix}, \code{dgCMatrix}, \code{matrix}, \code{pcalg}s \code{amat},
or a data frame with \code{from}, \code{to}, and \code{edge_type} columns.
}
