% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi}
\alias{as_caugi}
\title{Convert to a \code{caugi_graph}}
\usage{
as_caugi(
  x,
  class = c("DAG", "PDAG", "PAG", "Unknown"),
  simple = TRUE,
  build = TRUE,
  collapse = FALSE,
  collapse_to = "---",
  ...
)
}
\arguments{
\item{x}{An object to convert to a \code{caugi_graph}.}

\item{class}{"DAG", "PDAG", "PAG", or "Unknown".
"PAG" is only supported for integer coded matrices.}

\item{simple}{logical. If \code{TRUE} (default) the graph will be simple
(no multiple edges or self-loops).}

\item{build}{logical. If \code{TRUE} (default) build the graph now, otherwise
build lazily on first query or when using \code{\link[=build]{build()}}.}

\item{collapse}{logical. If \code{TRUE} collapse mutual directed edges to
undirected edges. Default is \code{FALSE}.}

\item{collapse_to}{Character string to use as the edge glyph when collapsing.
Should be a registered symmetrical edge glyph. Default is \code{"---"}.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A \code{caugi_graph} object.
}
\description{
Convert an object to a \code{caugi_graph}. The object can be a
\code{graphNEL}, ``matrix\verb{, or a data frame with }from`, `to`, and
`edge_type` columns.
}
\details{
For matrices, \code{as_caugi} assumes that the rows are the \code{from} nodes
and the columns are the \code{to} nodes. Thus, for a graph, G: A --> B, we would
have that \code{G["A", "B"] == 1} and \code{G["B", "A"] == 0}.
For PAGs, the integer codes are as follows (as used in \code{pcalg}):
\itemize{
\item 0: no edge
\item 1: tail (e.g., \verb{A o-- B} or \code{A --- B})
\item 2: arrowhead (e.g., \verb{A --> B} or \verb{A o-> B})
\item 3: circle (e.g., \verb{A o-o B})
}
}
\seealso{
Other conversion: 
\code{\link{as_igraph}()}
}
\concept{conversion}
