% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_caugi.R
\name{as_caugi.dgCMatrix}
\alias{as_caugi.dgCMatrix}
\alias{as_caugi.sparseMatrix}
\title{Create a caugi_graph from a sparse \code{dgCMatrix} or \code{sparseMatrix}}
\usage{
\method{as_caugi}{dgCMatrix}(x, directed = TRUE, ...)

\method{as_caugi}{sparseMatrix}(x, directed = TRUE, ...)
}
\arguments{
\item{x}{A sparse matrix of class \code{dgCMatrix} or \code{sparseMatrix}.}

\item{directed}{Logical; if \code{TRUE} (default) the graph is directed.}

\item{...}{Additional arguments (will not be passed to anything).}
}
\description{
Create a caugi_graph from a sparse \code{dgCMatrix} or \code{sparseMatrix}
}
