% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_to.R
\name{as_dagitty}
\alias{as_dagitty}
\title{Convert a caugi_graph to a dagitty graph}
\usage{
as_dagitty(x)
}
\arguments{
\item{x}{A \code{caugi_graph} object.}
}
\value{
A \code{dagitty} object.
}
\description{
Convert a caugi_graph to a dagitty graph
}
\seealso{
Other conversion: 
\code{\link{as_adjacency}()},
\code{\link{as_bnlearn}()},
\code{\link{as_caugi}()},
\code{\link{as_igraph}()}
}
\concept{conversion}
