% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.caugi_graph}
\title{Convert a caugi_graph to an igraph object}
\usage{
as_igraph(x, ...)

\method{as_igraph}{caugi_graph}(x, ...)
}
\arguments{
\item{x}{A caugi_graph}

\item{...}{Passed on to igraph::graph_from_data_frame()}

\item{directed}{Logical; if \code{NULL} (the default) we auto-detect any directed edges.}
}
\value{
An igraph object with an \code{edge_type} attribute on each edge.
}
\description{
Convert a caugi_graph to an igraph object
}
\seealso{
igraph::graph_from_data_frame
}
