% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert a caugi_graph to an igraph object}
\usage{
as_igraph(x, ...)
}
\arguments{
\item{x}{A \code{caugi_graph} object.}

\item{...}{Additional arguments passed to \code{igraph::graph_from_data_frame()}.}
}
\value{
An \code{igraph} object representing the same graph structure.
}
\description{
Convert a caugi_graph to an igraph object
}
\seealso{
Other conversion: 
\code{\link{as_caugi}()}
}
\concept{conversion}
