% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{build}
\alias{build}
\title{Build the \code{caugi} graph}
\usage{
build(cg)
}
\arguments{
\item{cg}{A \code{caugi} object.}
}
\value{
The input \code{caugi} object, invisibly.
}
\description{
DEPRECATED. This function is no longer necessary as graph sessions are always created lazily.
This function is now a no-op that returns the input graph invisibly. It is kept for backward compatibility,
but it will be removed in a future release. You can safely remove calls to \code{build()} from your code.
}
\examples{
cg <- caugi(A \%-->\% B, class = "DAG")
build(cg)

}
\seealso{
Other verbs: 
\code{\link{caugi_verbs}}
}
\concept{verbs}
