% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{build}
\alias{build}
\title{Build the \code{caugi} graph}
\usage{
build(cg)
}
\arguments{
\item{cg}{A \code{caugi} object.}
}
\value{
The input \code{caugi} object (invisibly), with its graph built.
}
\description{
Forces lazy compilation of the \code{caugi} graph
without running a specific query. Useful to pre-initialize the graph.
}
\examples{
cg <- caugi(A \%-->\% B, class = "DAG")
build(cg) # initialize graph without querying

}
\seealso{
Other verbs: 
\code{\link{caugi_verbs}}
}
\concept{verbs}
