% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi-package.R
\docType{package}
\name{caugi-package}
\alias{caugi-package}
\title{\code{caugi} (Causal Graph Interface)}
\description{
Create, query, and modify causal graphs. 'caugi' (Causal Graph Interface) is
a causality-first, high performance graph package that provides a simple
interface to build, structure, and examine causal relationships.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://caugi.org/}
  \item \url{https://github.com/frederikfabriciusbjerre/caugi}
  \item Report bugs at \url{https://github.com/frederikfabriciusbjerre/caugi/issues}
}

}
\author{
\strong{Maintainer}: Frederik Fabricius-Bjerre \email{frederik@fabriciusbjerre.dk} (\href{https://orcid.org/0009-0004-1079-1616}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Frederik Fabricius-Bjerre \email{frederik@fabriciusbjerre.dk} (\href{https://orcid.org/0009-0004-1079-1616}{ORCID}) [copyright holder]
  \item Johan Larsson \email{johan@jolars.co} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})
  \item Michael Sachs \email{sachsmc@gmail.com} (\href{https://orcid.org/0000-0002-1279-8676}{ORCID})
  \item Bjarke Hautop Kristensen \email{bjarke.hautop@gmail.com}
}

Other contributors:
\itemize{
  \item Leonard Henckel (author of the 'gadjid' crate, from which src/rust/src/graph/aid.rs is derived (MPL-2.0); see LICENSE.note) [contributor, copyright holder]
  \item Theo Würtzen (author of the 'gadjid' crate; see LICENSE.note) [contributor, copyright holder]
  \item Sebastian Weichwald (author of the 'gadjid' crate; see LICENSE.note) [contributor, copyright holder]
  \item Authors of the bundled Rust crates (vendored under src/rust/vendor; see LICENSE.note) [contributor, copyright holder]
}

}
\keyword{internal}
\keyword{package}
