% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi-package.R
\docType{package}
\name{caugi-package}
\alias{caugi}
\alias{caugi-package}
\title{caugi: Causal Graph Interface}
\description{
A tidy‐verse‐style causal graph package backed by a C++ CSR.
}
\author{
\strong{Maintainer}: Frederik Fabricius-Bjerre \email{frederik@fabriciusbjerre.dk}

}
\keyword{internal}
