% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi-package.R
\docType{package}
\name{caugi-package}
\alias{caugi-package}
\title{caugi: A tidy interface to causal graphs}
\description{
\strong{caugi} wraps a high-performance Rust core in a pipe-friendly R API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frederikfabriciusbjerre.github.io/caugi/}
  \item Report bugs at \url{https://github.com/frederikfabriciusbjerre/caugi/issues}
}

}
\author{
Frederik Fabricius-Bjerre \email{frederik@fabriciusbjerre.dk}
}
\keyword{internal}
\keyword{package}
