% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{caugi_export}
\alias{caugi_export}
\title{S7 Base Class for Caugi Exports}
\usage{
caugi_export(content = character(0), format = character(0))
}
\arguments{
\item{content}{A character string containing the exported graph.}

\item{format}{A character string indicating the export format.}
}
\description{
A base class for all caugi export formats. Provides common structure
and behavior for different export formats (DOT, GraphML, etc.).
}
\seealso{
Other export: 
\code{\link{caugi_deserialize}()},
\code{\link{caugi_dot}()},
\code{\link{caugi_graphml}()},
\code{\link{caugi_mermaid}()},
\code{\link{caugi_serialize}()},
\code{\link{export-classes}},
\code{\link{format-caugi}},
\code{\link{format-dot}},
\code{\link{format-graphml}},
\code{\link{format-mermaid}},
\code{\link{knit_print.caugi_export}},
\code{\link{read_caugi}()},
\code{\link{read_graphml}()},
\code{\link{to_dot}()},
\code{\link{to_graphml}()},
\code{\link{to_mermaid}()},
\code{\link{write_caugi}()},
\code{\link{write_dot}()},
\code{\link{write_graphml}()},
\code{\link{write_mermaid}()}
}
\concept{export}
