% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{caugi_from_adj_matrix}
\alias{caugi_from_adj_matrix}
\title{Create a caugi_graph from an adjacency matrix}
\usage{
caugi_from_adj_matrix(mat, directed = TRUE)
}
\arguments{
\item{mat}{A binary adjacency matrix (0/1) or a sparseMatrix}
}
\description{
Create a caugi_graph from an adjacency matrix
}
