% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-graphml.R
\name{caugi_graphml}
\alias{caugi_graphml}
\title{S7 Class for GraphML Export}
\usage{
caugi_graphml(content)
}
\arguments{
\item{content}{A character string containing the GraphML format graph.}
}
\value{
A \code{caugi_graphml} S7 object wrapping the GraphML string.
}
\description{
An S7 object that wraps a GraphML format string for caugi graphs.
}
\seealso{
Other export:
\code{\link[=caugi_deserialize]{caugi_deserialize()}},
\code{\link[=caugi_dot]{caugi_dot()}},
\code{\link[=caugi_export]{caugi_export()}},
\code{\link[=caugi_mermaid]{caugi_mermaid()}},
\code{\link[=caugi_serialize]{caugi_serialize()}},
\code{\link{export-classes}},
\code{\link{format-caugi}},
\code{\link{format-dot}},
\code{\link{format-graphml}},
\code{\link{format-mermaid}},
\code{\link[=knit_print.caugi_export]{knit_print.caugi_export()}},
\code{\link[=read_caugi]{read_caugi()}},
\code{\link[=read_graphml]{read_graphml()}},
\code{\link[=to_dot]{to_dot()}},
\code{\link[=to_graphml]{to_graphml()}},
\code{\link[=to_mermaid]{to_mermaid()}},
\code{\link[=write_caugi]{write_caugi()}},
\code{\link[=write_dot]{write_dot()}},
\code{\link[=write_graphml]{write_graphml()}},
\code{\link[=write_mermaid]{write_mermaid()}}
}
\concept{export}
