% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-graphml.R
\name{caugi_graphml}
\alias{caugi_graphml}
\title{S7 Class for GraphML Export}
\usage{
caugi_graphml(content)
}
\arguments{
\item{content}{A character string containing the GraphML format graph.}
}
\description{
An S7 object that wraps a GraphML format string for caugi graphs.
}
\seealso{
Other export: 
\code{\link{caugi_deserialize}()},
\code{\link{caugi_dot}()},
\code{\link{caugi_export}()},
\code{\link{caugi_mermaid}()},
\code{\link{caugi_serialize}()},
\code{\link{export-classes}},
\code{\link{format-caugi}},
\code{\link{format-dot}},
\code{\link{format-graphml}},
\code{\link{format-mermaid}},
\code{\link{knit_print.caugi_export}},
\code{\link{read_caugi}()},
\code{\link{read_graphml}()},
\code{\link{to_dot}()},
\code{\link{to_graphml}()},
\code{\link{to_mermaid}()},
\code{\link{write_caugi}()},
\code{\link{write_dot}()},
\code{\link{write_graphml}()},
\code{\link{write_mermaid}()}
}
\concept{export}
