% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{caugi_layout_circle}
\alias{caugi_layout_circle}
\title{Circle Layout}
\usage{
caugi_layout_circle(x, ...)
}
\arguments{
\item{x}{A \code{caugi} object.}

\item{...}{Ignored. For future extensibility.}
}
\value{
A \code{data.frame} with columns \code{name}, \code{x}, and \code{y} containing node
names and their coordinates.
}
\description{
Computes node coordinates by placing nodes evenly along the perimeter of a
circle. The first node is placed at the top of the circle, and subsequent
nodes proceed counter-clockwise. Edge structure is ignored. Works with all
edge types and produces deterministic results.
}
\examples{
cg <- caugi(
  A \%-->\% B,
  B \%-->\% C,
  C \%-->\% D,
  D \%-->\% A
)
layout <- caugi_layout_circle(cg)

}
\seealso{
Other plotting: 
\code{\link{add-caugi_plot-caugi_plot}},
\code{\link{caugi_layout}()},
\code{\link{caugi_layout_bipartite}()},
\code{\link{caugi_layout_fruchterman_reingold}()},
\code{\link{caugi_layout_kamada_kawai}()},
\code{\link{caugi_layout_sugiyama}()},
\code{\link{caugi_layout_tiered}()},
\code{\link{caugi_plot}()},
\code{\link{divide-caugi_plot-caugi_plot}},
\code{\link{plot}()}
}
\concept{plotting}
