% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-mermaid.R
\name{caugi_mermaid}
\alias{caugi_mermaid}
\title{S7 Class for Mermaid Export}
\usage{
caugi_mermaid(content)
}
\arguments{
\item{content}{A character string containing the Mermaid format graph.}
}
\description{
An S7 object that wraps a Mermaid format string for displaying caugi graphs.
When printed interactively, displays the Mermaid string cleanly.
}
\seealso{
Other export: 
\code{\link{caugi_deserialize}()},
\code{\link{caugi_dot}()},
\code{\link{caugi_export}()},
\code{\link{caugi_serialize}()},
\code{\link{export-classes}},
\code{\link{format-caugi}},
\code{\link{format-dot}},
\code{\link{format-mermaid}},
\code{\link{knit_print.caugi_export}},
\code{\link{read_caugi}()},
\code{\link{to_dot}()},
\code{\link{to_mermaid}()},
\code{\link{write_caugi}()},
\code{\link{write_dot}()},
\code{\link{write_mermaid}()}
}
\concept{export}
