% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{caugi_registry}
\alias{caugi_registry}
\title{Access the global edge registry, creating it if needed.}
\usage{
caugi_registry()
}
\value{
An \code{edge_registry} external pointer.
}
\description{
Access the global edge registry, creating it if needed.
}
\seealso{
Other registry: 
\code{\link{register_caugi_edge}()},
\code{\link{reset_caugi_registry}()},
\code{\link{seal_caugi_registry}()}
}
\concept{registry}
