% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_directed_PAG_to_symmetric_PAG}
\alias{collapse_directed_PAG_to_symmetric_PAG}
\title{Collapse mutual directed PAG edges (o->, not -->) to a single o-o edge}
\usage{
collapse_directed_PAG_to_symmetric_PAG(df)
}
\arguments{
\item{df}{A tibble with columns from, to, edge_type}

\item{collapse_to}{The code to use for collapsed directed pairs (default "---")}
}
\value{
A tibble with directed pairs collapsed
}
\description{
Collapse mutual directed PAG edges (o->, not -->) to a single o-o edge
}
\keyword{internal}
