% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_directed_to_symmetric_edges}
\alias{collapse_directed_to_symmetric_edges}
\title{Collapse mutual directed edges (-->) to a single symmetric edge (--- or <->)}
\usage{
collapse_directed_to_symmetric_edges(df, collapse_to = "---")
}
\arguments{
\item{df}{Tibble with columns \code{from}, \code{to}, \code{edge_type}}

\item{collapse_to}{Code to use for collapsed pairs (default "---")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Tibble with collapsed edge list
}\if{html}{\out{</div>}}
}
\description{
Collapse mutual directed edges (-->) to a single symmetric edge (--- or <->)
}
\keyword{internal}
