% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_symmetric_edges}
\alias{collapse_symmetric_edges}
\title{Collapse bidirected (<->) and undirected (---) duplicates}
\usage{
collapse_symmetric_edges(df)
}
\arguments{
\item{df}{Tibble with columns \code{from}, \code{to}, \code{edge_type}}
}
\value{
Tibble with duplicates removed
}
\description{
Collapse bidirected (<->) and undirected (---) duplicates
}
\keyword{internal}
