% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{condition_marginalize}
\alias{condition_marginalize}
\title{Marginalize and/or condition on variables in a DAG}
\usage{
condition_marginalize(cg, condvars = NULL, margvars = NULL)
}
\arguments{
\item{cg}{A caugi graph of class "DAG"}

\item{condvars}{Character vector of nodes to condition on}

\item{margvars}{Character vector of nodes to marginalize over}
}
\value{
A \code{caugi} object of class "DAG", "ADMG", or "UNKNOWN", depending on what type of edges are in the result
}
\description{
Marginalize variables out of a DAG, and/or condition on variables.
Depending on the structure, it could produce a graph with directed, bidirected, and undirected edges.
}
\examples{

mg <- caugi(U \%-->\% X + Y,
  A \%-->\% X,
  B \%-->\% Y,
  class = "DAG"
)

condition_marginalize(mg, margvars = "U") # ADMG
condition_marginalize(mg, condvars = "U") # DAG

}
\references{
Definition 4.2.1 in Thomas Richardson. Peter Spirtes. "Ancestral graph Markov models." Ann. Statist. 30 (4) 962 - 1030, August 2002. \url{https://doi.org/10.1214/aos/1031689015}
}
\seealso{
Other operations: 
\code{\link{exogenize}()},
\code{\link{latent_project}()},
\code{\link{moralize}()},
\code{\link{mutate_caugi}()},
\code{\link{skeleton}()}
}
\concept{operations}
