% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{dag_from_pdag}
\alias{dag_from_pdag}
\title{Extend a PDAG to a DAG using the Dor-Tarsi Algorithm}
\usage{
dag_from_pdag(PDAG)
}
\arguments{
\item{PDAG}{A \code{caugi} object of class \code{"PDAG"}.}
}
\value{
A \code{caugi} object of class \code{"DAG"} representing a DAG extension
of the input PDAG.
}
\description{
Given a Partially Directed Acyclic Graph (PDAG), this function attempts to
extend it to a Directed Acyclic Graph (DAG) by orienting the undirected edges
while preserving acyclicity and all existing directed edges.
The procedure implements the Dor-Tarsi algorithm.

If the PDAG cannot be consistently extended to a DAG, the function will raise an error.
}
\examples{
PDAG <- caugi(
  A \%---\% B,
  B \%---\% C,
  class = "PDAG"
)
DAG <- dag_from_pdag(PDAG)
edges(DAG)

}
\references{
Dor, D., & Tarsi, M. (1992). "A simple algorithm to construct a consistent
extension of a partially directed acyclic graph",
\url{https://api.semanticscholar.org/CorpusID:122949140}.
}
\seealso{
Other operations: 
\code{\link{condition_marginalize}()},
\code{\link{exogenize}()},
\code{\link{latent_project}()},
\code{\link{moralize}()},
\code{\link{mutate_caugi}()},
\code{\link{skeleton}()}
}
\concept{operations}
