% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{.cg_state}
\alias{.cg_state}
\title{Internal: Create the state environment for a \code{caugi_graph}}
\usage{
.cg_state(nodes, edges, ptr, built, simple, class)
}
\arguments{
\item{nodes}{A tibble of nodes with a \code{name} column.}

\item{edges}{A tibble of edges with \code{from}, \code{edge}, and \code{to} columns.}

\item{ptr}{A pointer to the underlying Rust graph structure
(or \code{NULL} if not built).}

\item{built}{Logical; whether the graph has been built.}

\item{simple}{Logical; whether the graph is simple
(no parallel edges or self-loops).}

\item{class}{Character; one of \code{"Unknown"}, \code{"DAG"}, or \code{"PDAG"}.}
}
\description{
Internal function to create the state environment for a
\code{caugi_graph}. This function is not intended to be used directly by users.
}
\keyword{internal}
