% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.relations}
\alias{.relations}
\title{Get relations of nodes in a \code{caugi_graph}}
\usage{
.relations(cg, nodes, getter)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{nodes}{A vector of node names or indices, a (vector of) unquoted
node name(s), or an expression combining these with \code{+} and \code{c()}.}

\item{getter}{A function that takes a pointer to the graph and a zero-based
node index, and returns a vector of zero-based indices of related nodes.}
}
\value{
A tibble with a \code{name} column.
}
\description{
Helper to get relations (parents, children, etc.) of nodes.
}
\keyword{internal}
