% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{.update_caugi}
\alias{.update_caugi}
\title{Update nodes and edges of a \code{caugi}}
\usage{
.update_caugi(
  cg,
  nodes = NULL,
  edges = NULL,
  action = c("add", "remove"),
  inplace = FALSE
)
}
\arguments{
\item{cg}{A \code{caugi} object.}

\item{nodes}{A tibble with column \code{name} for node names to add/remove.}

\item{edges}{A tibble with columns \code{from}, \code{edge}, \code{to} for edges to
add/remove.}

\item{action}{One of \code{"add"} or \code{"remove"}.}

\item{inplace}{Logical, whether to modify the graph inplace or not.}
}
\value{
The updated \code{caugi} object.
}
\description{
Internal helper to add or remove nodes/edges and mark graph as
not built.
}
\keyword{internal}
