% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{edge_types}
\alias{edge_types}
\title{Get the edge types of a \code{caugi_graph}.}
\usage{
edge_types(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
A character vector of edge types.
}
\description{
Get the edge types of a \code{caugi_graph}.
}
