% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{edges}
\alias{edges}
\alias{E}
\title{Get edges of a \code{caugi_graph}.}
\usage{
edges(cg)

E(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\description{
Get edges of a \code{caugi_graph}.
}
\seealso{
Other queries: 
\code{\link{ancestors}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{edge_types}()},
\code{\link{exogenous}()},
\code{\link{is_acyclic}()},
\code{\link{is_caugi}()},
\code{\link{is_cpdag}()},
\code{\link{is_dag}()},
\code{\link{is_empty_caugi}()},
\code{\link{is_pdag}()},
\code{\link{markov_blanket}()},
\code{\link{neighbors}()},
\code{\link{nodes}()},
\code{\link{parents}()},
\code{\link{same_nodes}()},
\code{\link{subgraph}()}
}
\concept{queries}
