% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{enumerate_dags}
\alias{enumerate_dags}
\title{Enumerate all DAGs in a Markov equivalence class}
\usage{
enumerate_dags(cg)
}
\arguments{
\item{cg}{A \code{caugi} object of class \code{"PDAG"} or \code{"MPDAG"}.}
}
\value{
A list of \code{caugi} objects of class \code{"DAG"}, one per DAG in the MEC.
}
\description{
Given a PDAG, enumerate every Directed Acyclic Graph (DAG) in its
Markov equivalence class -- i.e. every acyclic orientation of the undirected
edges that introduces no new v-structures.
}
\details{
Implements the recursive listing algorithm of Chickering (2002). The input
is normalized via \code{\link[=meek_closure]{meek_closure()}} so any background-knowledge orientations
are preserved and propagated before enumeration. The number of returned
DAGs can grow super-exponentially in the size of the chain components;
use \code{\link[=count_dags]{count_dags()}} first to size the problem.
}
\examples{
pdag <- caugi(
  A \%---\% B,
  B \%---\% C,
  class = "PDAG"
)
dags <- enumerate_dags(pdag)
length(dags) # 3

}
\references{
Chickering, D. M. (2002). "Learning Equivalence Classes of Bayesian-Network
Structures". \emph{Journal of Machine Learning Research}, 2:445--498.
}
\seealso{
Other operations: 
\code{\link{condition_marginalize}()},
\code{\link{count_dags}()},
\code{\link{dag_from_pdag}()},
\code{\link{exogenize}()},
\code{\link{latent_project}()},
\code{\link{meek_closure}()},
\code{\link{moralize}()},
\code{\link{mutate_caugi}()},
\code{\link{normalize_latent_structure}()},
\code{\link{skeleton}()}
}
\concept{operations}
