% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{exogenous}
\alias{exogenous}
\alias{exo}
\title{Get all exogenous nodes in a \code{caugi_graph}}
\usage{
exogenous(cg, undirected_as_parents = FALSE)

exo(cg, undirected_as_parents = FALSE)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{undirected_as_parents}{Logical; if \code{TRUE}, undirected edges are treated
as (possible) parents, if \code{FALSE} (default), undirected edges are ignored.}
}
\value{
A tibble with a \code{name} column.
}
\description{
Get all exogenous nodes (nodes with no parents) in a
\code{caugi_graph}.
}
