% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-graphml.R
\name{format-graphml}
\alias{format-graphml}
\title{GraphML Format Export and Import}
\description{
Functions for converting caugi graphs to and from GraphML format.
GraphML is an XML-based file format for graphs supported by many
graph tools and libraries.
}
\seealso{
Other export:
\code{\link[=caugi_deserialize]{caugi_deserialize()}},
\code{\link[=caugi_dot]{caugi_dot()}},
\code{\link[=caugi_export]{caugi_export()}},
\code{\link[=caugi_graphml]{caugi_graphml()}},
\code{\link[=caugi_mermaid]{caugi_mermaid()}},
\code{\link[=caugi_serialize]{caugi_serialize()}},
\code{\link{export-classes}},
\code{\link{format-caugi}},
\code{\link{format-dot}},
\code{\link{format-mermaid}},
\code{\link[=knit_print.caugi_export]{knit_print.caugi_export()}},
\code{\link[=read_caugi]{read_caugi()}},
\code{\link[=read_graphml]{read_graphml()}},
\code{\link[=to_dot]{to_dot()}},
\code{\link[=to_graphml]{to_graphml()}},
\code{\link[=to_mermaid]{to_mermaid()}},
\code{\link[=write_caugi]{write_caugi()}},
\code{\link[=write_dot]{write_dot()}},
\code{\link[=write_graphml]{write_graphml()}},
\code{\link[=write_mermaid]{write_mermaid()}}
}
\concept{export}
