% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_graph}
\alias{generate_graph}
\title{Random caugi graph with exactly \code{n} nodes and \code{m} edges.}
\usage{
generate_graph(n, m = NULL, p = NULL, mode = c("DAG", "CPDAG"))
}
\arguments{
\item{n}{Integer >= 0. Number of nodes in the graph.}

\item{m}{Integer in \verb{0, n*(n-1)/2}. Number of edges in the graph. Exactly one
of \code{m} or \code{p} must be supplied.}

\item{p}{Numeric in \verb{[0,1]}. Probability of edge creation. Exactly one of
\code{m} or \code{p} must be supplied.}

\item{mode}{"DAG" or "CPDAG".}
}
\value{
The sampled \code{caugi_graph} object.
}
\description{
Sample a random DAG or CPDAG using
Erdős-Rényi for random graph generation.
}
\concept{simulation}
\concept{simulation functions}
