% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hd}
\alias{hd}
\title{Hamming Distance}
\usage{
hd(cg1, cg2, normalized = FALSE)
}
\arguments{
\item{cg1}{A \code{caugi_graph} object.}

\item{cg2}{A \code{caugi_graph} object.}

\item{normalized}{Logical; if \code{TRUE}, returns the normalized Hamming
Distance.}
}
\value{
An integer representing the Hamming Distance between the two graphs.
}
\description{
Compute the Hamming Distance between two graphs.
}
\seealso{
Other metrics: 
\code{\link{aid}()},
\code{\link{shd}()}
}
\concept{metrics}
