% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_ag}
\alias{is_ag}
\title{Is the \code{caugi} graph an AG?}
\usage{
is_ag(cg, force_check = FALSE)
}
\arguments{
\item{cg}{A \code{caugi} object.}

\item{force_check}{Logical; if \code{TRUE}, the function will test if the graph is
an AG, if \code{FALSE} (default), it will look at the graph class and match
it, if possible.}
}
\value{
A logical value indicating whether the graph is an AG.
}
\description{
Checks if the given \code{caugi} graph is an
Ancestral Graph (AG).

An AG contains directed (\verb{-->}), bidirected (\verb{<->}), and undirected (\verb{---})
edges, and must satisfy ancestral graph constraints (no directed cycles,
anterior constraint, and undirected constraint).
}
\examples{
cg_ag <- caugi(
  A \%-->\% B,
  C \%<->\% D,
  E \%---\% F,
  class = "AG"
)
is_ag(cg_ag) # TRUE

cg_ug <- caugi(
  A \%---\% B,
  class = "UG"
)
is_ag(cg_ug) # TRUE (UGs are valid AGs)

}
\seealso{
Other queries: 
\code{\link{ancestors}()},
\code{\link{anteriors}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{districts}()},
\code{\link{edge_types}()},
\code{\link{edges}()},
\code{\link{exogenous}()},
\code{\link{is_acyclic}()},
\code{\link{is_admg}()},
\code{\link{is_caugi}()},
\code{\link{is_cpdag}()},
\code{\link{is_dag}()},
\code{\link{is_empty_caugi}()},
\code{\link{is_mag}()},
\code{\link{is_pdag}()},
\code{\link{is_simple}()},
\code{\link{is_ug}()},
\code{\link{m_separated}()},
\code{\link{markov_blanket}()},
\code{\link{neighbors}()},
\code{\link{nodes}()},
\code{\link{parents}()},
\code{\link{posteriors}()},
\code{\link{same_nodes}()},
\code{\link{spouses}()},
\code{\link{subgraph}()},
\code{\link{topological_sort}()}
}
\concept{queries}
