% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_caugi}
\alias{is_caugi}
\title{Is it a \code{caugi} graph?}
\usage{
is_caugi(x, throw_error = FALSE)
}
\arguments{
\item{x}{An object to check.}

\item{throw_error}{Logical; if \code{TRUE}, throws an error if \code{x} is not a
\code{caugi_graph}.}
}
\value{
A logical value indicating whether the object is a \code{caugi_graph}.
}
\description{
Checks if the given object is a \code{caugi_graph}.
}
