% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_dag}
\alias{is_dag}
\title{Is the \code{caugi} graph a DAG?}
\usage{
is_dag(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
A logical value indicating whether the graph is a DAG.
}
\description{
Checks if the given \code{caugi} graph is a
Directed Acyclic Graph (DAG).
}
