% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_empty_caugi}
\alias{is_empty_caugi}
\title{Is the \code{caugi} graph empty?}
\usage{
is_empty_caugi(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
A logical value indicating whether the graph is empty.
}
\description{
Checks if the given \code{caugi} graph is empty (has no nodes).
}
