% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_empty_caugi}
\alias{is_empty_caugi}
\title{Is the \code{caugi} graph empty?}
\usage{
is_empty_caugi(cg)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
A logical value indicating whether the graph is empty.
}
\description{
Checks if the given \code{caugi} graph is empty (has no nodes).
}
\seealso{
Other queries: 
\code{\link{ancestors}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{edge_types}()},
\code{\link{edges}()},
\code{\link{exogenous}()},
\code{\link{is_acyclic}()},
\code{\link{is_caugi}()},
\code{\link{is_dag}()},
\code{\link{is_pdag}()},
\code{\link{markov_blanket}()},
\code{\link{neighbors}()},
\code{\link{nodes}()},
\code{\link{parents}()},
\code{\link{same_nodes}()}
}
\concept{queries}
