% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_mpdag}
\alias{is_mpdag}
\title{Is the \code{caugi} graph an MPDAG?}
\usage{
is_mpdag(cg)
}
\arguments{
\item{cg}{A \code{caugi} object.}
}
\value{
A logical value indicating whether the graph is an MPDAG.
}
\description{
Checks if the given \code{caugi} graph is a
Maximally oriented Partially Directed Acyclic Graph
(MPDAG), i.e. a PDAG where no additional edge orientations
are implied by Meek's rules (R1--R4).
}
\details{
If the graph is not PDAG-compatible, the function returns \code{FALSE}.
}
\examples{
cg_not_mpdag <- caugi(
  A \%---\% B,
  A \%-->\% C,
  C \%-->\% B,
  class = "PDAG"
)
is_mpdag(cg_not_mpdag) # FALSE

}
\seealso{
Other queries: 
\code{\link{ancestors}()},
\code{\link{anteriors}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{districts}()},
\code{\link{edge_types}()},
\code{\link{edges}()},
\code{\link{exogenous}()},
\code{\link{is_acyclic}()},
\code{\link{is_admg}()},
\code{\link{is_ag}()},
\code{\link{is_caugi}()},
\code{\link{is_cpdag}()},
\code{\link{is_dag}()},
\code{\link{is_empty_caugi}()},
\code{\link{is_mag}()},
\code{\link{is_pdag}()},
\code{\link{is_simple}()},
\code{\link{is_ug}()},
\code{\link{m_separated}()},
\code{\link{markov_blanket}()},
\code{\link{neighbors}()},
\code{\link{nodes}()},
\code{\link{parents}()},
\code{\link{posteriors}()},
\code{\link{same_nodes}()},
\code{\link{spouses}()},
\code{\link{subgraph}()},
\code{\link{topological_sort}()}
}
\concept{queries}
