% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_pdag}
\alias{is_pdag}
\title{Is the \code{caugi} graph a PDAG?}
\usage{
is_pdag(cg, test = FALSE)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
A logical value indicating whether the graph is a PDAG.
}
\description{
Checks if the given \code{caugi} graph is a
Partially Directed Acyclic Graph (PDAG).
}
