% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{is_valid_backdoor}
\alias{is_valid_backdoor}
\title{Is a backdoor set valid?}
\usage{
is_valid_backdoor(
  cg,
  X = NULL,
  Y = NULL,
  Z = NULL,
  X_index = NULL,
  Y_index = NULL,
  Z_index = NULL
)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{X, Y}{Single node names.}

\item{Z}{Optional node set for conditioning}

\item{X_index, Y_index, Z_index}{Optional 1-based indices.}
}
\value{
Logical scalar.
}
\description{
Checks whether \code{Z} is a valid backdoor adjustment set for
\verb{X --> Y}.
}
\seealso{
Other adjustment: 
\code{\link{adjustment_set}()},
\code{\link{all_backdoor_sets}()},
\code{\link{d_separated}()}
}
\concept{adjustment}
