% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length-caugi}
\alias{length-caugi}
\alias{length}
\title{Length of a \code{caugi}}
\usage{
length(x)
}
\arguments{
\item{x}{A \code{caugi} object.}
}
\value{
An integer representing the number of nodes.
}
\description{
Returns the number of nodes in the graph.
}
\examples{
cg <- caugi(
  A \%-->\% B,
  class = "DAG"
)
length(cg) # 2

cg2 <- caugi(
  A \%-->\% B + C,
  nodes = LETTERS[1:5],
  class = "DAG"
)
length(cg2) # 5

}
\seealso{
Other caugi methods:
\code{\link{caugi-equality}},
\code{\link[=print]{print()}}
}
\concept{caugi methods}
\concept{methods}
