% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length}
\alias{length}
\title{Length of a \code{caugi_graph}}
\arguments{
\item{x}{A \code{caugi_graph} object.}
}
\value{
An integer representing the number of nodes.
}
\description{
Returns the number of nodes in the graph.
}
\examples{
cg <- caugi_graph(
  A \%-->\% B,
  class = "DAG"
)
length(cg) # 2

cg2 <- caugi_graph(
  A \%-->\% B + C,
  nodes = LETTERS[1:5],
  class = "DAG"
)
length(cg2) # 5

}
\seealso{
Other caugi_graph methods: 
\code{\link{print}()}
}
\concept{caugi_graph methods}
\concept{methods}
