% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_graph.R
\name{length}
\alias{length}
\title{Length of a \code{caugi_graph}}
\arguments{
\item{cg}{A \code{caugi_graph} object.}
}
\value{
An integer representing the number of nodes.
}
\description{
Returns the number of nodes in the graph.
}
