% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{meek_closure}
\alias{meek_closure}
\title{Apply Meek closure to a PDAG}
\usage{
meek_closure(cg)
}
\arguments{
\item{cg}{A \code{caugi} object. Must be PDAG-compatible.}
}
\value{
A \code{caugi} object of class \code{"PDAG"} that is closed under Meek's
rules.
}
\description{
Applies Meek's orientation rules (R1--R4) repeatedly to a PDAG until no more
orientations are implied.
}
\examples{
pdag <- caugi(
  A \%---\% B,
  A \%-->\% C,
  C \%-->\% B,
  class = "PDAG"
)
mpdag <- meek_closure(pdag)
edges(mpdag)

}
\seealso{
Other operations: 
\code{\link{condition_marginalize}()},
\code{\link{dag_from_pdag}()},
\code{\link{exogenize}()},
\code{\link{latent_project}()},
\code{\link{moralize}()},
\code{\link{mutate_caugi}()},
\code{\link{normalize_latent_structure}()},
\code{\link{skeleton}()}
}
\concept{operations}
