% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{mutate_caugi}
\alias{mutate_caugi}
\title{Mutate \code{caugi} class}
\usage{
mutate_caugi(cg, new_class)
}
\arguments{
\item{cg}{A \code{caugi} object.}

\item{new_class}{A character string specifying the new class.}
}
\value{
A \code{caugi} object of the specified class.
}
\description{
Mutate the \code{caugi} class from one graph class to another, if possible.
For example, convert a \code{DAG} to a \code{PDAG}, or a fully directed \code{caugi} of
class \code{UNKNOWN} to a \code{DAG}. Throws an error if not possible.
}
