% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{neighbors}
\alias{neighbors}
\alias{neighbours}
\alias{nb}
\alias{neighborhood}
\alias{neighbourhood}
\title{Get neighbors of nodes in a \code{caugi_graph}}
\usage{
neighbors(cg, nodes)

neighbours(cg, nodes)

nb(cg, nodes)

neighborhood(cg, nodes)

neighbourhood(cg, nodes)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{nodes}{A vector of node names or indices, a (vector of) unquoted
node name(s), or an expression combining these with \code{+} and \code{c()}.}
}
\description{
Get neighbors of nodes in a \code{caugi_graph}
}
