% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{parents}
\alias{parents}
\title{Get parents of nodes in a \code{caugi_graph}}
\usage{
parents(cg, nodes = NULL, index = NULL)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{nodes}{A vector of node names, a vector of unquoted
node names, or an expression combining these with \code{+} and \code{c()}.}

\item{index}{A vector of node indexes.}
}
\description{
Get parents of node in a graph. Note that not both nodes and index can be
given.
}
