% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{parents}
\alias{parents}
\alias{pa}
\title{Get parents of nodes in a \code{caugi_graph}}
\usage{
parents(cg, nodes)

pa(cg, nodes)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{nodes}{A vector of node names or indices, a (vector of) unquoted
node name(s), or an expression combining these with \code{+} and \code{c()}.}
}
\description{
Get parents of nodes in a \code{caugi_graph}
}
