% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\title{Print a \code{caugi}}
\usage{
print(x, max_nodes = getOption("caugi.max_nodes"),
  max_edges = getOption("caugi.max_edges"), ...)
}
\arguments{
\item{x}{A \code{caugi} object.}

\item{max_nodes}{Optional numeric; maximum number of node names to consider.
If \code{NULL}, the method automatically prints as many as fit on one console
line (plus a separate truncation line if needed).}

\item{max_edges}{Optional numeric; maximum number of edges to consider.
If \code{NULL}, the method automatically prints as many edges as fit on two
console lines (plus a separate truncation line if needed).}

\item{...}{Not used.}
}
\value{
The input \code{caugi} object, invisibly.
}
\description{
Print a \code{caugi}
}
\examples{
cg <- caugi(A \%-->\% B, class = "DAG")
print(cg)

}
\seealso{
Other caugi methods:
\code{\link{caugi-equality}},
\code{\link{length-caugi}}
}
\concept{caugi methods}
\concept{methods}
