% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\title{Print a \code{caugi}}
\arguments{
\item{x}{A \code{caugi} object.}

\item{...}{Not used.}
}
\value{
The input \code{caugi} object, invisibly.
}
\description{
Print a \code{caugi}
}
\examples{
cg <- caugi(
  A \%-->\% B + C,
  nodes = LETTERS[1:5],
  class = "DAG"
)
print(cg)

}
\seealso{
Other caugi methods: 
\code{\link{length}()}
}
\concept{caugi methods}
\concept{methods}
