% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\title{Print a \code{caugi_graph}}
\arguments{
\item{x}{A \code{caugi_graph} object.}

\item{...}{Not used.}
}
\value{
The input \code{caugi_graph} object, invisibly.
}
\description{
Print a \code{caugi_graph}
}
\examples{
cg <- caugi_graph(
  A \%-->\% B + C,
  nodes = LETTERS[1:5],
  class = "DAG"
)
print(cg)

}
\seealso{
Other caugi_graph methods: 
\code{\link{length}()}
}
\concept{caugi_graph methods}
\concept{methods}
