% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_caugi_edge}
\alias{register_caugi_edge}
\title{Register a new edge type in the global registry.}
\usage{
register_caugi_edge(
  glyph,
  left_mark,
  right_mark,
  orientation,
  class,
  symmetric = FALSE,
  traversable_when_conditioned = TRUE
)
}
\arguments{
\item{glyph}{A string representing the edge glyph (e.g., \code{"-->"}, \code{"<->"}).}

\item{left_mark}{A string representing the left mark type.
Possible values: "line", "circle", "arrow", "other".}

\item{right_mark}{A string representing the right mark type.
Possible values: "line", "circle", "arrow", "other".}

\item{orientation}{A string representing the edge orientation.
Possible values: "none", "right_head", "left_head", "both_heads".}

\item{class}{A string representing the edge class.
Possible values: "directed", "undirected", "bidirected", "partial".}

\item{symmetric}{Logical; if TRUE, the edge is considered symmetric.}

\item{traversable_when_conditioned}{Logical; if TRUE, the edge can be
traversed, when conditioning on the nodes that it connects.}
}
\value{
The integer code assigned to the registered edge type.
}
\description{
This function allows you to register a new edge type (glyph)
in the global edge registry used by \code{caugi_graph()}.
}
