% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_caugi_edge}
\alias{register_caugi_edge}
\title{Register a new edge type in the global registry.}
\usage{
register_caugi_edge(
  glyph,
  tail_mark,
  head_mark,
  class,
  symmetric = FALSE,
  flags = NULL
)
}
\arguments{
\item{glyph}{A string representing the edge glyph (e.g., \code{"-->"}, \code{"<->"}).}

\item{tail_mark}{One of "arrow", "tail", "circle", "other".}

\item{head_mark}{One of "arrow", "tail", "circle", "other".}

\item{class}{One of "directed","undirected","bidirected","partial".}

\item{symmetric}{Logical.}

\item{flags}{A character vector of flags. Currently supported:
\itemize{
\item "TRAVERSABLE_WHEN_CONDITIONED"
\item "LATENT_CONFOUNDING"
}}
}
\value{
The integer code assigned to the registered edge type.
}
\description{
Register a new edge type in the global registry.
}
