% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaaa-coding-helpers.R
\name{roclet_output.roclet_cargo_test}
\alias{roclet_output.roclet_cargo_test}
\title{Output for the cargo tests roclet}
\usage{
roclet_output.roclet_cargo_test(x, results, base_path, ...)
}
\arguments{
\item{x}{The roclet object.}

\item{results}{The results from \code{roclet_process}.}

\item{base_path}{The base path of the package.}

\item{...}{Additional arguments (not used).}
}
\description{
This function runs \verb{cargo test} in the Rust source directory if a
\code{Cargo.toml} file is found there.
}
