% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaaa-coding-helpers.R
\name{roclet_process.roclet_cargo_test}
\alias{roclet_process.roclet_cargo_test}
\title{Process for the cargo tests roclet}
\usage{
roclet_process.roclet_cargo_test(x, blocks, env, base_path)
}
\arguments{
\item{x}{The roclet object.}

\item{blocks}{The parsed roxygen blocks.}

\item{env}{The environment in which to evaluate the blocks.}

\item{base_path}{The base path of the package.}
}
\description{
This function is a no-op because the actual work is done in the
\code{roclet_output}.
}
