% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{same_nodes}
\alias{same_nodes}
\title{Same nodes?}
\usage{
same_nodes(cg1, cg2, throw_error = FALSE)
}
\arguments{
\item{cg1}{A \code{caugi_graph} object.}

\item{cg2}{A \code{caugi_graph} object.}

\item{throw_error}{Logical; if \code{TRUE}, throws an error if the graphs do not
have the same nodes.}
}
\value{
A logical indicating if the two graphs have the same nodes.
}
\description{
Check if two \code{caugi_graph} objects have the same nodes.
}
\seealso{
Other queries: 
\code{\link{ancestors}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{edge_types}()},
\code{\link{edges}()},
\code{\link{exogenous}()},
\code{\link{is_acyclic}()},
\code{\link{is_caugi}()},
\code{\link{is_dag}()},
\code{\link{is_empty_caugi}()},
\code{\link{is_pdag}()},
\code{\link{markov_blanket}()},
\code{\link{neighbors}()},
\code{\link{nodes}()},
\code{\link{parents}()},
\code{\link{subgraph}()}
}
\concept{queries}
