% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{subgraph}
\alias{subgraph}
\title{Extract a subgraph induced by a set of nodes}
\usage{
subgraph(cg, ...)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{...}{Unquoted node names, vectors via \code{c()}, or \code{+} composition.}
}
\value{
A \code{caugi_graph} object containing only the specified nodes and any
edges between them.
}
\description{
Extract a subgraph induced by a set of nodes.
}
\seealso{
Other queries: 
\code{\link{ancestors}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{edge_types}()},
\code{\link{edges}()},
\code{\link{exogenous}()},
\code{\link{is_acyclic}()},
\code{\link{is_caugi}()},
\code{\link{is_dag}()},
\code{\link{is_empty_caugi}()},
\code{\link{is_pdag}()},
\code{\link{markov_blanket}()},
\code{\link{neighbors}()},
\code{\link{nodes}()},
\code{\link{parents}()},
\code{\link{same_nodes}()}
}
\concept{queries}
