% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{subgraph}
\alias{subgraph}
\title{Extract a subgraph induced by a set of nodes}
\usage{
subgraph(cg, ...)
}
\arguments{
\item{cg}{A \code{caugi_graph} object.}

\item{...}{Unquoted node names, vectors via \code{c()}, or \code{+} composition.}
}
\value{
A \code{caugi_graph} object containing only the specified nodes and any
edges between them.
}
\description{
Extract a subgraph induced by a set of nodes.
}
